renames the OrderLegCollection class to better represent order legs
